Update mprLib.c
Generally this API is called to generate XSRF Token with size=32.
bytes is allocated for 16 chars.
Now clal to mprGetRandomBytes is made to fill those 16 chars with random numbers between 0 to 255. However, size being passed is just sizeof(bytes) which returns 4 (size of a pointer) and not 16 which is length of string. So only first 4 chars are randomnly generated. Remaining characters are whichver left from the mprAlloc call.